Lots of updates and fixes. Started process of cleaning up unittests so they make more sense and document the apis. Simplified some examples to convey their point more clearly. Few fixes to make clang not complain. Updated to several submodules. Switched to using std::type_index instead of type_info due to guarantees promised by the standard. Fully typed systems are now returned from add_system(), which allows the compiler to inline the system update call.

+++ benchmark/ecs.cpp
@@ -17,7 +17,7 @@ struct shared_s : ecs::shared {
 };
 
 auto constexpr benchmark_system = [](ecs::entity_id ent, int& color, shared_s const& shared) {
-	constexpr size_t max_iterations = 500;
+	constexpr int max_iterations = 500;
 	constexpr double fr_w = 1.5;
 	constexpr double fr_h = 1.5;
 	constexpr double fr_x = -2.2;
